Updated file structure to seperate certain classes.
