+++ GPU TEM-STEM Simulation/STEMAreaDialog.xaml.cs
@@ -131,10 +131,6 @@ namespace GPUTEMSTEMSimulation
         }
     }
 
-
-
-
-
     public class StemAreaArgs : EventArgs
     {
         private STEMArea msg;
@@ -146,31 +142,5 @@ namespace GPUTEMSTEMSimulation
         {
             get { return msg; }
         }
-    }
-}
-
-
-public class STEMArea
-{
-    public float xStart { get; set; }
-
-    public float xFinish { get; set; }
-
-    public float yStart { get; set; }
-
-    public float yFinish { get; set; }
-
-    public int xPixels { get; set; }
-
-    public int yPixels { get; set; }
-
-    public float getxInterval
-    {
-        get { return (xFinish - xStart) / xPixels; } // maybe abs
-    }
-
-    public float getyInterval
-    {
-        get { return (yFinish - yStart) / yPixels; }
     }
 }
