+++ GPU TEM-STEM Simulation/Utils/MainWindowUtils.cs
